Struct pill_core::pill_typemap::PillTypeMap

pub struct PillTypeMap(HashMap<TypeId, Box<dyn Any + Send>>);

PillTypeMap is a simple abstraction around the standard library’s HashMap type, where types are its keys. This allows for statically-checked value retrieval.

Implementations

impl PillTypeMap

pub fn new() -> Self

Creates a new instance of PillTypeMap.

pub fn contains_key<T>(&self) -> bool where
T: PillTypeMapKey,

Returns true if the map contains a value for the specified PillTypeMapKey.

use pill_typemap::{PillTypeMap, PillTypeMapKey};

struct Number;

impl PillTypeMapKey for Number {
    type Storage = i32;
}

let mut map = PillTypeMap::new();
assert!(!map.contains_key::<Number>());
map.insert::<Number>(42);
assert!(map.contains_key::<Number>());

pub fn insert<T>(&mut self, value: T::Storage) where
T: PillTypeMapKey,

Inserts a new value based on its PillTypeMapKey. If the value has been already inserted, it will be overwritten with the new value.

use pill_typemap::{PillTypeMap, PillTypeMapKey};

struct Number;

impl PillTypeMapKey for Number {
    type Storage = i32;
}

let mut map = PillTypeMap::new();
map.insert::<Number>(42);
// Overwrite the value of `Number` with -42.
map.insert::<Number>(-42);

pub fn entry<T>(&mut self) -> Entry<'_, T> where
T: PillTypeMapKey,

Retrieve the entry based on its PillTypeMapKey

pub fn get<T>(&self) -> Option<&T::Storage> where
T: PillTypeMapKey,

Retrieve a reference to a value based on its PillTypeMapKey. Returns None if it couldn’t be found.

use pill_typemap::{PillTypeMap, PillTypeMapKey};

struct Number;

impl PillTypeMapKey for Number {
    type Storage = i32;
}

let mut map = PillTypeMap::new();
map.insert::<Number>(42);

assert_eq!(*map.get::<Number>().unwrap(), 42);

pub fn get_mut<T>(&mut self) -> Option<&mut T::Storage> where
T: PillTypeMapKey,

Retrieve a mutable reference to a value based on its PillTypeMapKey. Returns None if it couldn’t be found.

use pill_typemap::{PillTypeMap, PillTypeMapKey};

struct Number;

impl PillTypeMapKey for Number {
    type Storage = i32;
}

let mut map = PillTypeMap::new();
map.insert::<Number>(42);

assert_eq!(*map.get::<Number>().unwrap(), 42);
*map.get_mut::<Number>().unwrap() -= 42;
assert_eq!(*map.get::<Number>().unwrap(), 0);

pub fn remove<T>(&mut self) -> Option<T::Storage> where
T: PillTypeMapKey,

Removes a value from the map based on its PillTypeMapKey, returning the value or None if the key has not been in the map.

use pill_typemap::{PillTypeMap, PillTypeMapKey};

struct Text;

impl PillTypeMapKey for Text {
    type Storage = String;
}

let mut map = PillTypeMap::new();
map.insert::<Text>(String::from("Hello PillTypeMap!"));
assert!(map.remove::<Text>().is_some());
assert!(map.get::<Text>().is_none());